requirements.txt
requests
environs~=8.0.0
python-telegram-bot